Remove PEERS message type (#152)
When a node connects to a tracker, it requests a random list of nodes and then proceeds to connect to those. This is no longer needed as a node will connect to those nodes that share responsibility for same streams. These peers are requested from the tracker via STREAM_INFO messages.

Removed the code related to fetching random peers from tracker after connecting.