Optimization: do not propagate to sender (#162)
A node receiving a valid message (i.e. not a duplicate) will remove the sender of the message from the list of nodes to propagate the message further to. There is no obvious point in sending the message back to the source we already received it from.

Included is an integration test that sets up a network of 3 fully-connected nodes, sends a single message, waits for the message to propagate through the network, and then checks that the total number of duplicate messages was 2 (instead of 4).