Allow previous number to be null (in any state) of DuplicateMessageDetector (#194)
Method `DuplicateMessageDetector#markAndCheck(previousNumber, number)` is used to mark a message as seen and to check whether the message has been seen before.

The assumption before was that only the first message in a streamId-producerId session would have `previousNumber = null`, the messages thereafter having it always set to refer to the previous message. This assumption was too strict, we should now assume that `previousNumber = null` at any point in time.

I updated `DuplicateMessageDetector` to work with this assumption. It provides primitive duplicate detection (based on last messageId seen) when checking a message that does not have previous number set.