Node: handle failing sends in propagation (#223)
Node may fail to propagate data to other nodes if for example sockets are in closing state. Worst case scenario (that comes up in tests but not not necessarily that often in reality) is that we fail to forward message to any node.

Added logic that puts message back to buffer if it was not propagated to enough nodes. Later on another propagation of the message will be re-tried.