StreamFetcher: log errors when failing to contact E&E (#38)
A very typical reason for Data API to not work properly in an environment is its failure to communicate with E&E (e.g. given E&E url wrong or E&E not running)

Added logging to console when E&E communication fails so its easier to diagnose the problem.