fix bug in MicroBatchingStrategy (#75)
This PR fixes a bug that causes some messages not be stored in Cassandra.

Previously a batch was only cycled when it became full or _after_ it had been successfully written to Cassandra. The problem was that messages could be pushed to batches whilst  _in the middle_ of the process of writing to Cassandra. This caused the leftover messages to not be necessarily stored at all. 

Fixed this by adding a simple boolean field `this.committed` that is set to true immediately before we commence writing to Cassandra. This boolean is checked (in addition to checking whether a batch is full) when deciding whether it is time to cycle the batch.

Also added a small test for this scenario.