Add validation to message constructors (#24)
Add validation to the constructors of message classes. Two benefits. Firstly, allows us to catch errors caused by missing fields earlier in the execution of a request. Secondly, allows us to get rid of some basic validation checks in client and broker as we can assume that presence (and types) of fields have been checked in constructor. 

### Changeset
- 1st commit: validation functions and unit tests for them.
- 2nd commit: adding (and replacing existing) validation into messages classes. I'm not entirely sure about the optionality of some of the fields, especially those of the older versions V28, V29, and V0. Thus I opted to assume optionality in cases in which I was unsure. 
- 3rd commit: this is a change to the behavior of `toOtherVersion` which can potentially break something if we are depending on this behavior somewhere. It is quite arbitrary, so I'm wondering if this is the case really.
4th commit: Fix a few tests, e.g., some were using integers as msgChainId. Some fixes are related to the behavior change introduced in the 3rd commit.

### Version upgrade
This PR has a high chance of breaking something in broker or client. Thus we should treat it as a _major_ version upgrade and take necessary precautions before migrating broker or client to use this changeset.