Instruction Throttling (#428)
In this PR the logic of instruction handling on node side is changed as follows.

1. Node maintains an instance of `InstructionThrottler` during its lifecycle.
2. Upon receiving an instruction from tracker, it calls `InstructionThrottler#add` to enqueue the instruction.
3. `InstructionThrottler` invokes a callback supplied by Node (i.e. `Node#handleTrackerInstruction`) whenever a new instruction should be handled.

The class `InstructionThrottler` makes sure that no more than one instruction is handled at a time. While an instruction is being handled and a new instruction arrives, it enqueues the new instruction. If several instructions with the same streamId arrive during handling, it will keep only the latest one around (basically the one with the highest counter).

## Some considerations
- This does add complexity to the instruction handling logic. It would be nice if we could somehow check if this change brings more stability to the network.
-  A previous version of this change had a less strict invariant. It allowed for more than one instruction to be handled at a time given that they operate on different streams. (So basically max 1 instruction under handling at a time per streamId.)