NET-59: Decode forward slashes ("/") in streamId URL paths (#472)
Started by writing tests to ensure that streamIDs with forward slashes are properly handled in tracker HTTP endpoints. Noticed that it doesn't work but that `decodeURIComponent` is needed. Ended up writing more tests for other endpoints and noticed that they don't properly respond with an error... so ended up fixing rest of endpoints.