NET-124: Typescript tests (#496)
Converted tests to TypeScript. All tests pass, but some should be refactored later.

Created separate tsconfig.json configs for development (e.g. running tests) and building the distribution package. With this setup the test code is excluded from the distribution package.

Added TypeScript support to eslint, but disabled some rules so that the tests pass the validation. In NET-125 we'll define a good lint rule set, and most likely we could re-enable some of the rules which we disabled here.

Suggested refactoring:

- Tests should not use private fields (NET-153)
- Refine the start logic of startNetworkNode/startStorageNode (NET-154).