NET-32: Storage node config (#505)
New required option storageConfig for starting a storage node. The option contains an interface, which we can query the streams that the storage node should store. We can also add a listener for the config to notify the storage node about configuration changes.

With this change the storage node stores only the configured streams, not all streams. It subscribes to the configured streams and doesn't need the tracker to orchestrate the subscriptions.