Error handling of getStream/createStream/etc. (#208)
createStream, getStream, getStreamByName and getOrCreateStream reject the promise if there is an API error.

To support this functionality, a new errorCode field was added to AuthFetchError. Currently the ErrorCode enum lists the most typical API error codes: VALIDATION_ERROR and NOT_FOUND. Other codes can be added later, if needed.