NET-194: fix latency metrics (#514)
Latency metric was reported as a sum of latencies instead of last average latency. Fixed this by introducing a new metric type, "fixed" metric, in the metrics "framework". It is a metric that can be just simply set at any arbitrary point in time to any value and it will get reported as such.