NET-230: Fix resend range if fromTimestamp is before first bucket (#224)
Bug fix for fetching resend data when fromTimestamp is before first bucket. In that situation, a call to storage.requestRange responded with an empty result, and TypeError: Cannot read property 'dateCreate' of undefined. was written to the error log.

Fixed the bucket date range analysis to explicitly handle the case where only one bucket is found. Enhanced also the case where no messages are available (the same error message was written to the error logs in that case, too).

Added new integration tests and grouped the existing tests.