NET-86: Add a way to toggle on strict (Brubeck) validation (#59)
Currently StreamMessageValidator implements the validation rules present in the Corea network. In particular, it still allows:

unsigned data

unencrypted data in non-public streams

It would be good to be able to pass flags to StreamMessageValidator, which would enable stricter validation to be needed in Brubeck era:

Unsigned StreamMessages are invalid

Unencrypted StreamMessages are invalid unless the anonymous user has public stream_subscribe permission (stream is public)

Would be good if both conditions (signing & encryption) have separate flags. By default, the strict flags would be off by default for the time being. But in a future Brubeck testnet, we would toggle them on.