NET-248: Refactor adapters to plugins (#12)
Adapters are now plugins. Each plugin extends from Plugin class. HTTP Server is a shared service for all plugins, you can add HTTP endpoints by calling this.addHttpServerRouter() in a plugin. If a plugin has some configuration options, it should provide a schema for the configuration (by overriding this.getConfigSchema()).

Breaking change: The "adapters" section has been renamed to "plugins". That section is not an array, but a key-value object (the name attribute is now the key for the object). The config properties of plugins have not changed.