Rework reconnects, fix flaky tests (#562)
Mostly fixes flaky tests in CI.

Fast webRTC reconnect test is still sometimes fails in CI, but the occurances are rare. It appears that in rare cases a connection is closed just before the RemoteDescription is being set, this causes the connection forming process to stall. This could be fixed with a queuing mechanism, but expecting all such cases where the connection is closed and offer / answer signalling messages are received immediately after to wait for reconnects might lead to weird behaviour.

- Remove force connect logic
- Non-offerer always uses rtcConnect
- If connections have set RemoteDescription any new rtcConnect and rtcOffer messages are interpreted as reconnect attempts
- RemoteCandidate can no longer be called before RemoteDescription has been set
- Rework 'Messages are delivered on temporary loss of connection' test case