fix: clean up abi.ts
Only types are strictly required. This file should be kept small because it will be distributed with every StreamrClient to browsers etc.

The other methods are kept for now in case they might be needed in tests etc. in which case a separate ABI file should be produced for tests that has those functions uncommented.