Broker NET-333!: Configurable payload format for WebSocket and MQTT plugin (#105)
Unified the payload handling of websocket plugin and mqtt plugin. Added a new setting payloadMetadata (a boolean value) to both plugin configs.

If payloadMetadata is true, the publish and subscription payloads are JSONs which have a structure of
{
    content: ...,
    metadata: ...
}

For publish requests, it is possible to set an optional timestamp value in the metadata. For subscription requests, the metadata contains timestamp, sequenceNumber, publisherId and msgChainId fields of the message.

If payloadMetadata setting is false, the payloads are plain content JSONs (and therefore it is not possible to define the timestamp or read the metadata of the message).

Test refactoring: As the payload handling of websocket and mqtt plugins are similar, the plugin integration tests now use a shared helper method createMessagePluginTest. MQTT-specific test of streamIdDomain has been refactored to a separate unit test.

Breaking change: As the payloadFormat setting is a required parameter, this is be a breaking change.