broker: move legacy websocket plugin to use ws instead of uws (NET-323) (#75)
Due to the corrupted messages issue we were observing, and after trying numerous tricks to fix the uWS implementation, decided to change the library into `ws` after which the issue disappeared. Not sure yet what was actually causing the message corruption issue: there is reason to believe it could be originating deeper from within the Cassandra driver or our (mis-)use of it.

Included new configuration file _packages/broker/configs/development-prod-resend.env.json_ that allows one to hook up to the production storage node, which can be handy for testing purposes.