network [NET-436]: custom speedometer (#166)
Replaced the speedometer package with a custom utility. The utility is used to analyze the throughput rate of various metrics in MetricsContext and PerStreamMetrics.

The returned throughput rate results of the custom implementation are more accurate compared to the previous solution. Also, the new implementation doesn't need NodeJS.Timeout timers.

Removed also the redundant PerStreamMetrics class.