network [NET-472]: debounced instruction send for Tracker (#192)
Tracker instructions are collected to buffers and sent after a short delay. For each stream there is a separate buffer.

We use debouncing to delay the sending. We send all the instructions in a stream buffer, when either when of these conditions is satisfied:

- the topology stabilizes: no new instructions has been formed for the stream in X milliseconds
- the buffer times out: we have buffered an instruction for Y milliseconds

Added new command line options for the two configuration options. The default values are currently 100 ms and 2000 ms (we can optimize the values later, if needed).