fix broker [NET-493]: don't store sec metrics (#193)
Bug fix for metrics stream storage node assignments.

Expected:
We should should not assign the sec metrics to a storage node (but min, hour and day should be assigned)

Actual:
All for metrics stream are assigned, including the sec stream

Reason:
The if statement didn't filter out anything as compared a string to a number with !== operator. The root cause was in ensureStreamsCreated method which provided the periodLength as a string.

Added also a unit test for MetricsPublisher and fixed some type annotations.