network: remove support for full status messages (#223)
Currently all status messages that nodes send include only one stream. In #174 we modified the sending logic in node, but didn't change the tracker behavior (to keep the backwards compatibility).

In this PR we drop the backwards compatibility. The trackers is no longer able to process status messages that contain multiple streams.