broker: function return types (#241)
Now all functions have return types. Removed eslint exception "@typescript-eslint/explicit-module-boundary-types": "warn" so that eslint will show an error if we miss a return type in the future.

Changed Todo types to actual types in Stream, StreamStateManager and DeleteExpiredCmd. Stream and StreamStateManager are now generic classes (the connection type is a generic type).