client: [NET-594]: validate client options (#305)
Validate the options JSON given by the user, and throw an error if there are invalid settings.

Check that there are no missing required properties, and that the JSON doesn't contain any unknown properties (most likely typos). Also validate the content of each field if the content can be validated easily (e.g. an Ethereum address or an URL). The validation is implemented using ajv JSON schema validator.

The validation requires that the auth.privateKey is a string and it validates the content with a regexp. Earlier we supported also BytesLike (an undocumented feature).